requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.5.0

# HTTP & Scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.40.0

# Scheduling
apscheduler>=3.10.0

# Database
sqlite-utils>=3.36.0

# AI / LLM
openai>=1.12.0

# Utilities
rich>=13.7.0          # Pretty console output
tenacity>=8.2.0       # Retry logic
